Reachability checking v0.
Draft and incomplete (but cute looking?) platform-to-platform
reachability checking.  The game "rules" (well, parameters) are all in
find_direct_edges, which finds out whether platform 1 could get to
platform 2 assuming everything works out perfectly.  Later this will
be refined to include "arc does not interpenetrate either platform"
and "does not interpenetrate other platforms or dynamic objects" and
so on.  Mostly it's an experiment.  At some point I'll probably give
in and do concrete A* style search but I want to see how far we can
get with qualitative and constraint based methods.

You can drag platforms around to see the reachability relation but
it's hard to read because I didn't take the time to draw arrowheads.
It just draws "from" sides of edges a little further left than "to"
sides.